requirements.txt
absl-py==0.7.1
aiohttp==3.6.2
aiovk==3.0.0
astor==0.8.0
async-timeout==3.0.1
asyncpg==0.18.3
attrs==19.1.0
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
multidict==4.5.2
numpy==1.17.0
opt-einsum==3.1.0
protobuf==3.9.1
pycparser==2.19
PyNaCl==1.3.0
python-Levenshtein==0.12.0
PyYAML==5.1.2
requests==2.22.0
scipy==1.3.1
selectolax==0.1.13
six==1.12.0
tensorboard==1.14.0
tensorflow==1.15.2
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
vk-api==11.5.0
websockets==8.1
Werkzeug==0.15.5
wrapt==1.11.2
yarl==1.3.0
